Fix lints, and re-factor ocfis and sandbox
