+++ CHANGES.txt
@@ -3,8 +3,14 @@ Changelog
 
 Version 0.5.0, unreleased
 
-* Replaced custom smooth filter for scipy's ndimage filter.
-* Re-factor datasets to avoid the basemap syntax in favor of a `bbox` ike syntax.
+* Replaced custom smooth filter for `scipy`'s `ndimage` filter.
+* Re-factor `datasets` to avoid the `basemap` syntax in favor of a `bbox` like syntax.
+* Re-factor `ff_tools` into a single `ocfis` module.
+* Moved `compass` to plotting.
+* Added `get_gebco15` function to extract gebco15 bathymetry.
+* Removed `teaching` and `smoothers` modules as they were here only for learning purposes.
+* Removed shapely dependency. Using matplotlib `Path` in `in_polygon` instead.
+* Many speed improvements via lazy imports and updates.
 
 Version 0.4.0, 27-Oct-2016.
