+++ MANIFEST.in
@@ -1,13 +1,4 @@
 include README.rst
-include LICENSE.txt
-include CHANGES.txt
-include oceans/*.py
+include *.txt
+recursive-include oceans *.py
 include oceans/colormaps/cmap_data/*.dat
-include oceans/colormaps/*.py
-include oceans/datasets/*.py
-include oceans/ff_tools/*.py
-include oceans/plotting/*.py
-include oceans/RPSstuff/*.py
-include oceans/sandbox/*.py
-include oceans/sw_extras/*.py
-include oceans/tests/*.py
