+++ oceans/RPSstuff/RPSstuff.py
@@ -1,25 +1,11 @@
-from __future__ import absolute_import, division
+# -*- coding: utf-8 -*-
+
+from __future__ import (absolute_import, division, print_function)
 
 from datetime import datetime
 
 import numpy as np
 import numpy.ma as ma
-
-
-__all__ = ['h2hms',
-           'hms2h',
-           'ms2hms',
-           'julian',
-           'jdrps2jdmat',
-           'jdmat2jdrps',
-           'gregorian',
-           's2hms',
-           'ss2',
-           # FIXME: STOPPED HERE
-           'near',
-           'angled',
-           'coast2bln',
-           'fixcoast']
 
 
 earth_radius = 6371.e3
