+++ oceans/RPSstuff/__init__.py
@@ -1,7 +1,40 @@
+# -*- coding: utf-8 -*-
+
 """
 Miscellaneous tools from Rich Signell's RPSstuff (matlab)
 =========================================================
 
 """
 
-from .RPSstuff import *
+from .RPSstuff import (
+    h2hms,
+    hms2h,
+    ms2hms,
+    julian,
+    jdrps2jdmat,
+    jdmat2jdrps,
+    gregorian,
+    s2hms,
+    ss2,
+    near,
+    angled,
+    coast2bln,
+    fixcoast,
+    )
+
+
+__all__ = [
+    'h2hms',
+    'hms2h',
+    'ms2hms',
+    'julian',
+    'jdrps2jdmat',
+    'jdmat2jdrps',
+    'gregorian',
+    's2hms',
+    'ss2',
+    'near',
+    'angled',
+    'coast2bln',
+    'fixcoast',
+    ]
