+++ oceans/colormaps/__init__.py
@@ -1,1 +1,8 @@
+ # -*- coding: utf-8 -*-
+
 from .colormaps import get_color, cm
+
+__all__ = [
+    'get_color',
+    'cm'
+    ]
