--- oceans/ff_tools/__init__.py
@@ -1,4 +1,0 @@
-from .ocfis import *
-from .synop import *
-from .ff_tools import *
-from .teaching import *
