+++ oceans/ocfis/__init__.py
@@ -1,0 +1,56 @@
+# -*- coding: utf-8 -*-
+
+from .synop import scaloa
+from .ocfis import (
+    bin_dates,
+    binave,
+    binavg,
+    complex_demodulation,
+    del_eta_del_x,
+    despike,
+    lagcorr,
+    mld,
+    pcaben,
+    series_spline,
+    spdir2uv,
+    spec_rot,
+    uv2spdir,
+    pol2cart,
+    cart2pol,
+    wrap_lon180,
+    wrap_lon360,
+    alphanum_key,
+    get_profile,
+    strip_mask,
+    shiftdim,
+    weim,
+    smoo2,
+    )
+
+
+__all__ = [
+    'bin_dates',
+    'binave',
+    'binavg',
+    'complex_demodulation',
+    'del_eta_del_x',
+    'despike',
+    'lagcorr',
+    'mld',
+    'pcaben',
+    'series_spline',
+    'spdir2uv',
+    'spec_rot',
+    'uv2spdir',
+    'pol2cart',
+    'cart2pol',
+    'wrap_lon180',
+    'wrap_lon360',
+    'alphanum_key',
+    'get_profile',
+    'strip_mask',
+    'shiftdim',
+    'scaloa',
+    'weim',
+    'smoo2'
+    ]
