+++ oceans/plotting/__init__.py
@@ -1,1 +1,18 @@
-from .plotting import *
+# -*- coding: utf-8 -*-
+
+from .plotting import (
+    stick_plot,
+    landmask,
+    level_colormap,
+    get_pointsxy,
+    EditPoints
+    )
+
+__all__ = [
+    'stick_plot',
+    'landmask',
+    'level_colormap',
+    'get_pointsxy',
+    'compass',
+    'EditPoints'
+    ]
