+++ oceans/sandbox/__init__.py
@@ -1,8 +1,13 @@
-"""
-Misc tools
-==========
+# -*- coding: utf-8 -*-
 
-"""
+from .lines import (
+    LineNormals2D,
+    LineCurvature2D,
+    inverse3
+    )
 
-from .fdatasets import *
-from .lines import *
+__all__ = [
+    'LineNormals2D',
+    'LineCurvature2D',
+    'inverse3'
+    ]
